Move task servlet tests to correct package.
